refactor: moves screen files to presentation directory
