+++ lib/common/navigation/coordinator.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:polkadex/common/navigation/routes.dart';
 import 'package:polkadex/common/utils/enums.dart';
 import 'package:polkadex/features/landing/presentation/screens/market_token_selection_screen.dart';
-import 'package:polkadex/features/trade/widgets/card_flip_widgett.dart';
+import 'package:polkadex/features/trade/presentation/widgets/card_flip_widgett.dart';
 
 abstract class Coordinator {
   static final _navigationKey = GlobalKey<NavigatorState>();
