+++ lib/features/trade/presentation/screens/coin_trade_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:polkadex/common/configs/app_config.dart';
 import 'package:polkadex/common/dummy_providers/app_chart_dummy_provider.dart';
 import 'package:polkadex/common/navigation/coordinator.dart';
-import 'package:polkadex/features/trade/widgets/card_flip_widgett.dart';
-import 'package:polkadex/features/trade/widgets/order_book_widget.dart';
+import 'package:polkadex/features/trade/presentation/widgets/card_flip_widgett.dart';
+import 'package:polkadex/features/trade/presentation/widgets/order_book_widget.dart';
 import 'package:polkadex/common/providers/bottom_navigation_provider.dart';
 import 'package:polkadex/common/utils/colors.dart';
 import 'package:polkadex/common/utils/enums.dart';
