+++ lib/features/trade/presentation/widgets/order_book_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:polkadex/common/configs/app_config.dart';
 import 'package:polkadex/features/landing/presentation/dialogs/trade_view_dialogs.dart';
 import 'package:polkadex/features/landing/presentation/widgets/order_book_chart_item.dart';
-import 'package:polkadex/features/trade/order_book_item_model.dart';
+import 'package:polkadex/features/trade/presentation/order_book_item_model.dart';
 import 'package:polkadex/common/utils/colors.dart';
 import 'package:polkadex/common/utils/enums.dart';
 import 'package:polkadex/common/utils/extensions.dart';
